T: replace BuildViewManager in build test with test implementation
It allows:
- avoiding `if (isUnitTestMode)` code in `CargoBuildManager.kt`
- dropping static field from `CargoBuildManager`
- moving test implementation from production into test source set
